requirements.txt
deta==1.0.1
fastapi==0.70.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
